requirements.txt
# Flask web framework
Flask==2.2.3

# Template rendering
Jinja2==3.1.2

# Basic requirements
pathlib==1.0.1
python-dotenv==1.0.0
logging==0.4.9.6 